Extract CvTestTrait
